+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Updated `glow` version to `0.14`.
 - Fixed an incorrect feature name `bind_vertex_array_support` was called as `vertex_array_support`.
 - Changed example image to sipi-4-2-07 (peppers).
+- Changed `AutoRenderer::initialize` and `Renderer::initialize` to `AutoRenderer::new`
+  and `Renderer::new`, and deprecated both `initialize` methods.